[Fix] 회원프로필 수정(Patch profile) http method 변경(PUT->PATCH) ,찜 목록(get wishs)  최신순 정렬,회원프로필 수정(Patch profile) updateSellerInfo mapper 추가 및 유효성 검사 추가

+++ server-main/src/main/java/com/laonworks/shop/api/controller/handler/auth/WithdrawalHandler.java
@@ -59,7 +59,10 @@ public class WithdrawalHandler extends BaseHandler {
                 return res;
             }
             String encryptedPassword = CryptoUtils.encryptPassword(plainPassword,uservo.salt);
-
+            if(!uservo.password.equals(encryptedPassword)){
+                res.setCode(ResultCode.InvalidPassword);
+                return res;
+            }
             if (userType ==UserType.User.getValue()) {
                 result =authMapper.deleteUserInfo(email);
